+++ lib/pay/stripe/billable.rb
@@ -102,7 +102,7 @@ module Pay
       end
 
       def subscribe(name: Pay.default_product_name, plan: Pay.default_plan_name, **options)
-        quantity = options.delete(:quantity) || 1
+        quantity = options.delete(:quantity)
         opts = {
           expand: ["pending_setup_intent", "latest_invoice.payment_intent", "latest_invoice.charge.invoice"],
           items: [plan: plan, quantity: quantity],